Added support for saucelabs
See the comment in run-tests.rb to use saucelabs.  
You'll need to copy the saucelabs.sample.yml to saucelabs.yml and add your username and key.
You'll also need to startup a saucelabs tunnel if you are capybara test runner against a local server.  This can be done with sauce_connect which is available from saucelabs.com